+++ cmd/pam-module/args_linux_test.go
@@ -1,4 +1,6 @@
-package pam
+//go:build linux
+
+package main
 
 import (
 	"strings"